pyproject.toml
[tool.poetry]
name = "pdf-parser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.8.1"
pymupdf4llm = "^0.0.17"
openai = "^1.47.0"
pikepdf = "^9.3.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
